check isfile isdir

+++ odabot/cli.py
@@ -940,10 +940,16 @@ def make_galaxy_tools(obj, dry_run, loop, force, pattern):
                                     for fileglob in globlist:
                                         addfiles = glob(fileglob, root_dir=wf_repo_dir, recursive=True)
                                         for fp in addfiles:
-                                            shutil.copytree(
-                                                os.path.join(wf_repo_dir, fp),
-                                                os.path.join(outd, fp)
-                                                )
+                                            if os.path.isfile(os.path.join(wf_repo_dir, fp)):
+                                                shutil.copyfile(
+                                                    os.path.join(wf_repo_dir, fp),
+                                                    os.path.join(outd, fp)
+                                                    )
+                                            else:
+                                                shutil.copytree(
+                                                    os.path.join(wf_repo_dir, fp),
+                                                    os.path.join(outd, fp)
+                                                    )
                                     
 
                             logger.info("Git status:\n" + sp.check_output(['git', 'status'], text=True))
